==> VI_data_collection/3_response_intercept_analysis/4_item_exploratory_analysis_1.1.Rmd <==
---
title: "item analysis"
author: "andrew demetriou"
---

```{r setup, include=FALSE}
library('here')            # file logistics
library('data.table')      # data manipulation
library('dplyr')           # data manipulation
library('tidyr')           # data manipulation
library('ggplot2')         # visualization
library('ggridges')        # joyplot visualization
library('shiny')           # interactive visualization
#library('Rtsne')           # tsne visualization

options(scipen=999)
theme_set(theme_minimal())
```

```{r}
# load dataset
load(here("VI_data_collection", "_data", "survey_response_data", "response_df.RDS"))

# create a working dataset
responses_dt <- as.data.table(response_df)

rm(response_df)
```


```{r}
# replace 'song' with 'item_ID' as column name
setDT(responses_dt)

# subset columns for values questions:
responses_dt <- responses_dt %>%
  select("participant_ID", "item_ID", "w", "f", 
         paste0("1_", seq(1, 10)),
         "c")

# convert to numeric
columns <- c(paste0("1_", seq(1, 10)), "c")
responses_dt[, (columns):= lapply(.SD, as.numeric), .SDcols = columns]

responses_dt <- responses_dt %>%
  rename(writer=w, 
         familiar = f, 
         POWER = `1_1`, 
         ACHIEVEMENT = `1_2`, 
         HEDONISM = `1_3`, 
         STIMULATION = `1_4`, 
         SELF = `1_5`, 
         UNIVERSALISM = `1_6`, 
         BENEVOLENCE = `1_7`, 
         TRADITION = `1_8`, 
         CONFORMITY = `1_9`, 
         SECURITY = `1_10`, 
         confidence = c
         ) %>%
  select(participant_ID, item_ID, 
                         POWER, ACHIEVEMENT, HEDONISM, 
                         STIMULATION, SELF, UNIVERSALISM, 
                         BENEVOLENCE, TRADITION, 
                         CONFORMITY, SECURITY, 
                         writer, familiar, confidence)


values<-c("POWER", "ACHIEVEMENT", "HEDONISM",  
          "STIMULATION", "SELF", "UNIVERSALISM", 
          "BENEVOLENCE", "TRADITION",  
          "CONFORMITY", "SECURITY")

rm(columns)
```

```{r}
count_df <- responses_dt %>%
  group_by(item_ID) %>%
  count()

mean(count_df$n)
median(count_df$n)

count_df %>%
  ggplot(aes(n)) +
  geom_histogram(bins=10, color = 'black', fill = 'grey')

rm(count_df)
```

```{r}
responses_dt %>%
  pivot_longer(cols = POWER:SECURITY) %>%
  ggplot(aes(x=value, group=name, col = name)) +
  geom_density()
```


Make a list of plots to display the value ratings for each song:

```{r}
item_plot_dt <- responses_dt

# convert to long format
item_plot_dt <- as.data.table(item_plot_dt) %>% 
  melt(., id.vars = c("participant_ID", "item_ID"), 
       measure.vars = c("ACHIEVEMENT","BENEVOLENCE", "CONFORMITY", "HEDONISM", "POWER", "SECURITY", "SELF", "STIMULATION", "TRADITION", "UNIVERSALISM"))


# make a character vector of all song IDs
items <- unique(item_plot_dt$item_ID)

# make a function to plot all the value ratings for each song
item_plot_function <- function(item_plot_dt, item){
  item_plot_dt %>% filter(., item_ID == item) %>%
    ggplot(., aes(x=value, y=variable, fill = 0.5 -abs(0.5 -stat(ecdf)))) +
    stat_density_ridges(geom="density_ridges_gradient", calc_ecdf=TRUE, show.legend=FALSE) +
    scale_fill_viridis_c(option="C") +
    theme_minimal() +
    xlab("")
}

# make a list of plots, one for each song
list_of_song_by_value_plots <- lapply(items, function(i) item_plot_function(item_plot_dt, i))

#name each plot after the song item it represents
names(list_of_song_by_value_plots) <- items

#list_of_song_by_value_plots[201]
rm(item_plot_function, item_plot_dt)
```

Make a plot of the confidence scores for each song:

```{r}
#compute median
median_confidence <- responses_dt %>% group_by(item_ID) %>% dplyr::summarize(median = median(confidence, na.rm=TRUE))

#plot histogram of individual song's confidence ratings
plot_confidence <- function(item){
  
  #item <- items[200]
  #median_confidence %>% filter(item_ID == item) %>% .$median
  
  responses_dt %>% filter(., item_ID == item) %>%
    ggplot(aes(confidence)) +
    geom_histogram(color = 'black', fill = 'grey', bins=10)+
    geom_vline(xintercept  = median_confidence %>% filter(., item_ID == item)%>% .$median, color = 'red') + 
    ylab("") +
    xlab("Self reported Confidence in Ratings") +
    theme(legend.position = 'none')
}

#make list of confidence histograms
list_of_confidence_plots <- lapply(items, plot_confidence)

#name each plot in list after the item they represent
names(list_of_confidence_plots) <- items

rm(median_confidence, plot_confidence)
```


Dynamic display of results:

```{r}
ui <- fluidPage(
  titlePanel("Distribution of Ratings per Value by Song"),
  
  # drop down menu to display the values
  selectInput('choice', 'select song', choice = items), 
  
  #plot song plot
  plotOutput('song_by_value_plots'),
  
  #plot confidence ratings
  plotOutput('confidence_plot')
  
)

server <- function(input, output, session){
  
  # write a function that receives the input
  # and then filters dataframe column
  display_summary <- function(dt) {
    dt %>% select(stat, item_ID, input$choice)
  }
  
  # render the plots; alpha distribution by n
  output$song_by_value_plots <- renderPlot({
      #note that plots are pre-computed
      list_of_song_by_value_plots[input$choice]
  })
  
  output$confidence_plot <- renderPlot({
    list_of_confidence_plots[input$choice]
  })
  
  
  }
shinyApp(ui=ui, server=server)
```
```{r}
rm(confidence_dt, list_of_confidence_plots, list_of_song_by_value_plots, median_confidence, ui, server, plot_confidence, cols, confidence_ratings, data_file_path, file_name, items, values)
```
