==> VI_data_collection/3_response_intercept_analysis/7_estimate_means_1.0.Rmd <==
---
title: "estimate means and confidence-weighted means"
author: "andrew demetriou"
---

```{r setup, include=FALSE}
library('here')            # file logistics
library('data.table')      # data frame logistics
library('tidyverse')       # code logistics

values <- c("POWER", "ACHIEVEMENT", "HEDONISM",  
            "STIMULATION", "SELF", "UNIVERSALISM", 
            "BENEVOLENCE", "TRADITION",  
            "CONFORMITY", "SECURITY")

# load and format datasets
response_dfs <- readRDS(file=here("VI_data_collection", "_data", "survey_response_data", "response_dfs.RDS"))
```

We have four datasets loaded into a single list:

From current survey:
*`responses_dt_c` is only complete responses from the current survey
*`responses_dt` is both all complete and all partial responses from current survey
*`responses_dt_p` is all complete responses from the pilot study
*`responses_dt_a` is all responses from current survey and pilot study


```{r}
mean_dfs_1 <- lapply(response_dfs[1:2], function(x){
  x %>%
  group_by(item_ID) %>%
  dplyr::summarise(
    across(
      -c(participant_ID, row_number, writer, familiar), mean))
})

mean_dfs_2 <- lapply(response_dfs[3:4], function(x){
  x %>%
  group_by(item_ID) %>%
  dplyr::summarise(
    across(
      -c(participant_ID, row_number), mean))
})

participant_mean_dfs <- c(mean_dfs_1, mean_dfs_2)

rm(mean_dfs_1, mean_dfs_2)
```

```{r}
saveRDS(participant_mean_dfs, here("VI_data_collection", "_data", "intermediary_data", "participant_mean_dfs.RDS"))
```

```{r}
weighted_mean = function(x, w, ..., na.rm=F){
  if(na.rm){
    keep = !is.na(x)&!is.na(w)
    w = w[keep]
    x = x[keep]
  }
  weighted.mean(x, w, ..., na.rm=F)
}
```


```{r}
response_df <- response_dfs[[1]]

response_df$weight <- response_df$confidence / 100

mean_df_1 <- response_df %>%
  group_by(item_ID) %>%
  dplyr::summarise(
    across(
      -c(participant_ID, row_number, writer, familiar), ~ weighted_mean(.x, weight, na.rm=T)))

mean_df_2 <- response_df %>%
  group_by(item_ID) %>%
  dplyr::summarise(
    across(
      -c(participant_ID, row_number, writer, familiar), mean))
```

```{r}
for(value in values){
  print(cor(mean_df_1[[value]], mean_df_2[[value]]))
}
```

```{r}
participant_confidence_weighted_dfs <- lapply(response_dfs[1:2], function(df){
  df$weight <- df$confidence / 100
  
  df <- df %>%
  group_by(item_ID) %>%
  dplyr::summarise(
    across(
      -c(participant_ID, row_number, writer, familiar), ~ weighted_mean(.x, weight, na.rm=T)))
  
  return(df)
})
```

```{r}
saveRDS(participant_confidence_weighted_dfs, here("VI_data_collection", "_data", "intermediary_data", "participant_confidence_weighted_dfs.RDS"))
```
